--- a/eventbus-jmh/src/main/java/module-info.java
+++ b/eventbus-jmh/src/main/java/module-info.java
@@ -3,20 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 open module net.minecraftforge.eventbus.jmh {
-    requires cpw.mods.modlauncher;
-    requires cpw.mods.securejarhandler;
-
-    requires org.objectweb.asm;
-    requires org.objectweb.asm.commons;
-    requires org.objectweb.asm.tree;
     requires net.minecraftforge.eventbus;
-    requires jopt.simple;
+    requires org.jspecify;
     requires jmh.core;
+    requires jdk.unsupported; // needed by JMH for Unsafe
 
-    requires static org.jetbrains.annotations;
-    requires static net.minecraftforge.eventbus.testjars;
-    requires net.minecraftforge.unsafe;
+    requires net.minecraftforge.eventbus.testjars;
 
-    provides cpw.mods.modlauncher.api.ITransformationService with
-        net.minecraftforge.eventbus.benchmarks.MockTransformerService;
+    exports net.minecraftforge.eventbus.benchmarks;
 }