--- a/eventbus-test/build.gradle
+++ b/eventbus-test/build.gradle
@@ -13,31 +13,28 @@ repositories {
 }
 
 java {
-    toolchain.languageVersion = JavaLanguageVersion.of(16)
+    toolchain.languageVersion = JavaLanguageVersion.of(21)
 }
 
 license {
     header = rootProject.file("LICENSE-header.txt")
+    newLine = false
 }
 
 dependencies {
-    testImplementation(rootProject)
-    testImplementation(projects.eventbusTestJar)
-    testImplementation(libs.junit.api)
-    testImplementation(libs.securemodules)
-    testImplementation(libs.modlauncher)
-    testImplementation(libs.unsafe)
-    testImplementation(libs.asm)
-    testImplementation(libs.asm.tree)
-    testImplementation(libs.jopt.simple)
-    testImplementation(libs.log4j.api)
-    testRuntimeOnly(libs.bundles.junit.runtime)
-    testCompileOnly(libs.nulls)
+    testImplementation rootProject
+    testImplementation projects.eventbusTestJar
+    testImplementation libs.junit.api
+    testImplementation libs.jspecify.annotations
+    testRuntimeOnly libs.bundles.junit.runtime
 }
 
 extraJavaModuleInfo {
     failOnMissingModuleInfo = false
-    automaticModule('net.sf.jopt-simple:jopt-simple', 'jopt.simple')
+}
+
+tasks.named('test', Test) {
+    useJUnitPlatform()
 }
 
 tasks.register('testAll', AggregateTest) {
@@ -45,23 +42,25 @@ tasks.register('testAll', AggregateTest) {
     output = rootProject.file('test_results.html')
 }
 
-VALID_VMS.each { javaVendor, javaVersions ->
-    javaVersions.each { javaVersion ->
-        def output = file("build/test-results/${javaVendor}-${javaVersion}/")
+((Map<String, List<Integer>>) ext.VALID_VMS).forEach { javaVendor, javaVersions ->
+    for (Integer javaVersion in javaVersions) {
+        var output = file("build/test-results/${javaVendor}-${javaVersion}/")
         output.mkdirs()
-        def task = tasks.register("test${javaVendor}${javaVersion}", Test) {
+        var task = tasks.register("test${javaVendor}${javaVersion}", Test) {
             useJUnitPlatform()
-            javaLauncher.set(javaToolchains.launcherFor {
-                it.vendor.set(JvmVendorSpec."${javaVendor.toUpperCase(Locale.ROOT)}" as JvmVendorSpec)
-                it.languageVersion.set(JavaLanguageVersion.of(javaVersion))
-                it.implementation.set(JvmImplementation.VENDOR_SPECIFIC)
-            })
+            javaLauncher = javaToolchains.launcherFor {
+                vendor = JvmVendorSpec."${javaVendor.toUpperCase(Locale.ROOT)}" as JvmVendorSpec
+                languageVersion = JavaLanguageVersion.of(javaVersion)
+                implementation = JvmImplementation.VENDOR_SPECIFIC
+            }
             reports.html.destination = file("build/test-reports/${javaVendor}-${javaVersion}/")
             reports.junitXml.destination = output
         }
-        testAll.inputs.dir(output)
-        testAll.dependsOn(task)
-        testAll.mustRunAfter(task)
+        tasks.named('testAll') {
+            inputs.dir(output)
+            dependsOn task
+            mustRunAfter task
+        }
     }
 }