--- a/gradle.properties
+++ b/gradle.properties
@@ -1,4 +1,4 @@
 org.gradle.caching=true
 #org.gradle.parallel=true
 org.gradle.configureondemand=true
-#org.gradle.configuration-cache=true
+org.gradle.configuration-cache=true