--- a/src/main/java/net/minecraftforge/eventbus/api/IEventBusInvokeDispatcher.java
+++ b/src/main/java/net/minecraftforge/eventbus/api/IEventBusInvokeDispatcher.java
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) Forge Development LLC
- * SPDX-License-Identifier: LGPL-2.1-only
- */
-package net.minecraftforge.eventbus.api;
-
-public interface IEventBusInvokeDispatcher {
-    void invoke(IEventListener listener, Event event);
-}