--- a/src/main/java/net/minecraftforge/eventbus/api/event/InheritableEvent.java
+++ b/src/main/java/net/minecraftforge/eventbus/api/event/InheritableEvent.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+package net.minecraftforge.eventbus.api.event;
+
+import net.minecraftforge.eventbus.internal.Event;
+
+public non-sealed interface InheritableEvent extends Event {}