--- a/src/main/java/net/minecraftforge/eventbus/api/event/MutableEvent.java
+++ b/src/main/java/net/minecraftforge/eventbus/api/event/MutableEvent.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+package net.minecraftforge.eventbus.api.event;
+
+import net.minecraftforge.eventbus.internal.Event;
+import net.minecraftforge.eventbus.internal.MutableEventInternals;
+
+public non-sealed abstract class MutableEvent extends MutableEventInternals implements Event {}