--- a/src/main/java/net/minecraftforge/eventbus/api/event/RecordEvent.java
+++ b/src/main/java/net/minecraftforge/eventbus/api/event/RecordEvent.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+package net.minecraftforge.eventbus.api.event;
+
+import net.minecraftforge.eventbus.internal.Event;
+
+/**
+ * For read-only or shallowly-immutable records.
+ */
+public non-sealed interface RecordEvent extends Event {}