--- a/src/main/java/net/minecraftforge/eventbus/api/event/characteristic/package-info.java
+++ b/src/main/java/net/minecraftforge/eventbus/api/event/characteristic/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+@NullMarked
+package net.minecraftforge.eventbus.api.event.characteristic;
+
+import org.jspecify.annotations.NullMarked;