This PR is a full rewrite of Forge's EventBus, a conclusion to the NovaBus experiments worked on-and-off for the past
year (first started in late February 2024), incorporating some of the requested API designs discussed on the Discord.

## Table of contents
<!-- TOC -->
  * [What's NovaBus?](#whats-novabus)
  * [Why a full rewrite?](#why-a-full-rewrite)
  * [List of changes](#list-of-changes)
  * [Removed features](#removed-features)
  * [WIP/Todo](#wiptodo)
  * [Compatibility](#compatibility)
  * [Performance techniques](#performance-techniques)
    * [Posting](#posting)
    * [Registration](#registration)
  * [Benchmarks](#benchmarks)
    * [Posting](#posting-1)
      * [Dozen](#dozen)
      * [Hundred](#hundred)
      * [Single](#single)
      * [Scaling](#scaling)
      * [Posting data](#posting-data)
    * [Registration](#registration-1)
      * [Registration data](#registration-data)
  * [Future work](#future-work)
  * [Looking far ahead](#looking-far-ahead)
<!-- TOC -->

## What's NovaBus?
The codename for my experiments with EventBus. Over the past year, I've been playing around with many different ideas
and designs with others, learning a lot about the JVM and Java along the way. A small subset have already been
backported into Forge's EventBus, such as #58, #63 and #65. But there's a lot that can't be done without breaking
changes, which is what this PR is for.

To be clear, the codename is mainly to avoid confusion and verbosity when comparing v6 and v7. When I say NovaBus here,
I mean the new EventBus v7 which this PR represents. The name of the EventBus project remains unchanged.

## Why a full rewrite?
A good portion of the experiments were focused on maintaining partial backwards-compatibility with the current EventBus
and not making too many breaking changes. By the time we were finished, there was huge implementation complexity and
only a handful of desired features and minor performance improvements to show for it...

It became apparent that more meaningful improvements (in every metric - performance, API, impl...) would only be
possible when freeing ourselves from the constraints of the current design. After further discussion on Discord, the
decision was made to drop various unwanted features and to go ahead with a full rewrite.

## List of changes
- Cancellation and monitoring are decoupled from the event instance
- Listeners can be `Consumer`, `Predicate`, or `ObjBooleanBiConsumer`, the latter two for cancellable events
    - For cancellable events, listeners can declare they always or never cancel the event, allowing for additional
      optimisations
- `CancellableEventBus extends EventBus` to avoid breaking changes for non-cancellable listeners relying on events that
  change from cancellable to non-cancellable
- `ListenerList` and `EventBus` have been merged. Each `Event` has its own `EventBus`. An `Event` can have multiple
  `EventBus` instances, but each `EventBus` can only have one `Event`. A group of `EventBus`es is called a `BusGroup`.
  Here's a rough mapping from old to new:
    - `ListenerList` + `EventBus` -> `EventBus`
    - `EventBus` -> `BusGroup`
    - `BusBuilder` -> `EventCharacteristic` interfaces
- Strong encapsulation, taking full advantage of the Java module system and sealed types to better organise the system
  and allow flexibility for future internal changes without the risk of breaking the API. Strong separation of API spec
  and implementation details.
    - Only the `api` packages are exposed, the rest is internal
    - API is split into three packages based on purpose:
        - `event` for creating events
        - `listener` for creating listeners
        - `bus` for registering listeners, posting events and managing buses
- Forward-thinking design which pre-emptively accounts for Valhalla value classes, null-restricted types, etc...
    - Allows for further performance improvements in the future as the JVM evolves, without breaking changes
    - jSpecify is used for enforcing null-restricted types today
- `Event` is now a sealed interface to support records and multiple inheritance
- Event characteristic annotations have been replaced with interfaces to allow for more granular control (no
  annotation restrictions/limitations)
    - Inheritance support is ~~both~~ opt-in ~~and opt-out~~
        - Opt-in is done by implementing the `InheritableEvent` interface
        - ~~Opt-out is done by annotating an event that extends `InheritableEvent` with `@MarkerEvent`~~
    - Phase tracking is now only available for the Monitor phase and is opted in with the `MonitorAware` interface
    - Events can declare they are only fired once (`EventCharacteristic.SelfDestructing`) to automatically dispose of
      themselves after being posted to save memory
- Each `EventBus` determines its functionality based on the `Event` it's associated with and its registered listeners
    - The event isn't cancellable? Skip checking if the event was cancelled
    - The event is cancellable, but all listeners are `Consumer`s? Can also skip checking if the event was cancelled
    - Whole specialised invokers are generated, rather than only the array of listeners
- Priority is now a `byte` instead of an `enum`, larger numbers run first, with 0 being the default. Decided on a `byte`
  instead of `int` because:
    - Needing more than a couple of hundred priorities is likely a sign of bad code (this is the main reason)
        - CSS `z-index: 999999` moment
    - Priority comparisons are simpler and faster, especially considering any boxing is more likely to be covered by the
      JVM's preloaded autoboxing cache than an Integer
    - It's more friendly to Valhalla value classes and memory usage
- `receiveCanceled` has been removed for normal listeners, monitoring listeners have it implicitly
- Bulk registration
    - Bulk registration of listeners uses `LambdaMetaFactory` instead of manual class generation and transforms
    - Attempting bulk registration on a class containing no listeners will throw an exception to prevent accidental
      registration of the wrong class
    - Supplying an instance of a class will now also register static methods. No more artificial restrictions when
      registering instance vs static methods
    - Much more comprehensive validation to avoid silent failures when library users make mistakes
- Inheritance support is aware of sealed hierarchies to reduce memory usage
    - If the event is `final` (such as for implementations of `RecordEvent`), the child list will be a shared empty list
    - If the event is `sealed`, the child list will be exactly sized based on the number of permitted subclasses
- Events can be self-posting, allowing for more concise code in some cases (`new MyEvent().post()`)
- `BusGroup` supports trimming the backing lists of all loaded `EventBus`es in the group to save memory

## Removed features
- `receiveCanceled` for normal listeners (use monitoring listeners instead)
- Generic events (add a `Class<?>` field/record component to your event or use event inheritance)
- Per-post listener filtering
- HasResult (use a nullable `Boolean` result field/record component instead, or make your own enum type)
- Phase tracking for anything other than the Monitor phase
- Inherited listeners in bulk registration

## WIP/Todo
This is stuff I'd like some help with and/or plan to do in the near future:
- [x] Nicer errors when doing bulk registration
- [x] Test for possible concurrency issues in `BusGroup`
- [ ] Integrate into Forge (see [MinecraftForge#10495](https://github.com/MinecraftForge/MinecraftForge/pull/10495) for a PoC)
- [ ] Better documentation

## Compatibility
This is a full rewrite that deserves a major version bump. The underlying design is different enough that most existing
code wouldn't work without changes even if the packages and classes had the same names. I decided that it's better to
do one big breaking change now than to annoy developers with a lot of spread-out, smaller breaking changes.

Due to the scale of change, it might be good to merge this as EventBus 7.0.0-beta.1 or similar, to allow a short
breaking change window just in case we find anything important that needs fixing while integrating into Forge.

We should make a separate branch for EventBus 6.x so that we can continue maintaining it.

## Performance techniques
Event dispatch is specialised based on each event's capabilities, its listeners' capabilities, the number of listeners
and available JVM features. Invokers are generated lazily and cached, with careful consideration for JVM performance
pitfalls. The code for posting hotpaths has been iterated on a lot overtime, driven by benchmark results.

### Posting
In the case of zero listeners, EventBus v7 is able to dynamically instruct the JVM that the event dispatch is no-op and
is optimised away entirely by the C2 JIT. The performance is the same as if the event was never instantiated or posted,
effectively resulting in zero overhead for unused events without requiring special handling from library users.

Events with a single listener are resolved to a direct method call, skipping any redundant cancellation checks and loops.

Depending on the size of the compiled method bytecode for lambdas, a small number of listeners can be preloaded into
trusted finals with a manually unrolled loop. This allows for additional optimisations by the JVM that typically
wouldn't be possible with a normal array of listeners, without eating up too much of the inlining budget.

Cancellation checks are only performed if the event is cancellable and at least one listener possibly cancels the event.

If all listeners are known to never cancel the event, we can treat it as if the event isn't cancellable until a
possibly cancelling listener is registered later.

If a listener always cancels, we skip remaining listeners at invoker build time, reducing memory accesses and allowing
for some of the size-specific optimisations mentioned above to still apply even when the number of listeners would
otherwise exceed it.

There are additional tricks involved, but I think you get the idea... ;)

### Registration
A lot less indirection is involved in registration now and the new design of registering directly to events removes the
need for TypeTools. Adding individual listeners is basically wrapping the lambda in an `EventListener` instance and
adding it to a synchronized collection... that's pretty much it.

Bulk registration is no longer an entirely separate system - it's now a frontend for the same system used for individual
registration. It takes a class, does some reflection to list its methods, performs some sanity checks, converts them to
direct lambdas with `LambdaMetaFactory` and then registers those as if they were individual listeners. The class
generation is all done for us by the JDK and is much more efficient. Still slower than individual registration because
it shifts a bunch of work that would otherwise be done at compile-time to runtime, but that might be something we can
improve with compile-time transformers in the future.

Inheritance support is handled differently to significantly reduce the amount of work involved in building the sorted
list for the invoker. The child's backing list is preloaded with the parent's listeners if the event is inheritable.
After the EventBus is built, the parent EventBus is notified of the new child and future listeners added to/removed from
parents propagate to their children. Sorting is done directly on the backing list, so resorting is relatively cheap.

## Benchmarks
EventBus v7 offers significant performance improvements across the board in like-for-like benchmarks, no transforms
needed. I encourage reviewers to also run the benchmarks themselves to provide additional data points.

### Posting
#### Dozen
<img alt="Bar chart comparing EventBus v6 NoLoader, EventBus v6 ModLauncher and NovaBus in posting 12 listeners per event, 3 events per post, across dynamic, lambda and static benchmark suites." src="https://github.com/user-attachments/assets/b26297b9-3802-448b-8c65-21c15ec3d764" />

| Benchmark     | EventBus v6 NoLoader   | EventBus v6 ModLauncher | EventBus v7 (NovaBus) |
|---------------|------------------------|-------------------------|-----------------------|
| Dynamic dozen | 158.569 ns/op ± 2.755  | 156.574 ns/op ± 11.053  | 103.117 ns/op ± 1.346 |
| Lambda dozen  | 140.146 ns/op ± 7.550  | 119.641 ns/op ± 0.839   | 95.778 ns/op ± 2.403  |
| Static dozen  | 137.157 ns/op ± 18.634 | 125.666 ns/op ± 22.419  | 95.705 ns/op ± 0.854  |

Compared to EventBus v6 in ModLauncher mode, NovaBus is about 35% faster for dynamic listeners, 20% faster for lambda and
25% faster for static listeners. The improvements are of course even larger when comparing both in NoLoader mode.

#### Hundred
The numbers are less consistent when there's so many listeners, but NovaBus still manages to be within the same ballpark
of improvements as with a dozen listeners:

<img alt="Bar chart comparing the three but this time 100 listeners per event instead of a dozen" src="https://github.com/user-attachments/assets/d3f56d1c-f455-4d16-b2d7-348d4e4bb608" />

| Benchmark       | EventBus v6 NoLoader     | EventBus v6 ModLauncher  | EventBus v7 (NovaBus)    |
|-----------------|--------------------------|--------------------------|--------------------------|
| Dynamic hundred | 4710.419 ns/op ± 221.955 | 4281.470 ns/op ± 522.737 | 3420.696 ns/op ± 237.921 |
| Lambda hundred  | 4277.721 ns/op ± 71.710  | 3891.848 ns/op ± 772.864 | 3464.047 ns/op ± 195.707 |
| Static hundred  | 4344.508 ns/op ± 122.697 | 4036.750 ns/op ± 222.120 | 3393.881 ns/op ± 198.007 |

The relative performance difference between dynamic and static listeners has also reduced enough that I no longer
consider it worth thinking about from a performance perspective, especially combined with the gains over the previous
design.

#### Single
For single listeners, we can see NovaBus has allowed the JVM to optimise away the event dispatch entirely:

<img alt="Bar chart comparing the three but this time a single listener per event" src="https://github.com/user-attachments/assets/88ab75f8-c331-4bd1-bd95-db587e2ecbf0" />

| Benchmark      | EventBus v6 NoLoader  | EventBus v6 ModLauncher | EventBus v7 (NovaBus) |
|----------------|-----------------------|-------------------------|-----------------------|
| Dynamic single | 28.671 ns/op ± 0.239  | 17.236 ns/op ± 0.111    | 0.251 ns/op ± 0.026   |
| Lambda single  | 26.076 ns/op ± 0.313  | 17.118 ns/op ± 0.237    | 0.244 ns/op ± 0.014   |
| Static single  | 25.670 ns/op ± 0.306  | 16.472 ns/op ± 0.222    | 0.243 ns/op ± 0.011   |

This is intentional. The new design ensures a fully inlinable execution chain for upto a handful of listeners, which
allows the JVM to see that although a listener is registered, the listener method is no-op. The performance is the
same as an empty JMH benchmark method - it's as if the event was never instantiated and posted.

To ensure a like-for-like comparison, I did not change the existing benchmarks. There are real-world cases where people
accidentally register empty event listeners, so it's good to show that the new architecture can accommodate this.

I did look into solutions which avoid inlining, such as using JMH blackholes in all the events and consuming the event
instance from each listener, however this isn't straightforward to do in a way that reflects real-world usage with all
listeners not being no-ops. For instance, the blackhole approach prevented inlining of event instantiation for events
with constant values. Another approach was to use a `@CompilerControl` annotation to prevent inlining, but this only
worked for the initial event instantiation and not for the listener method calls.

We can reconsider the benchmarks in a future PR if needed. For reference, I got around 5ns/op for the single dynamic
listener case with the Blackhole approach and around 1ns/op with the CompilerControl approach, but I'm not confident
in the accuracy of these inlining mitigation numbers at this time and didn't apply those approaches to the other
benchmarks (and thus are not directly comparable).

#### Scaling
![Line chart comparing EventBus v6 NoLoader, EventBus v6 ModLauncher and NovaBus in posting various listeners per event, 3 events per post. Number of listeners on the horizontal axis, time on the vertical axis](https://github.com/user-attachments/assets/2890e0bd-8ddf-4669-9d36-2014581aca11)

![Line chart comparing EventBus v6 NoLoader, EventBus v6 ModLauncher and NovaBus in posting 0-5 inclusive listeners per event, 3 events per post. Number of listeners on the horizontal axis, time on the vertical axis](https://github.com/user-attachments/assets/ff416755-55d1-4966-810c-70ff1d157156)

![Line chart comparing EventBus v6 NoLoader, EventBus v6 ModLauncher and NovaBus in posting 10-50 inclusive listeners per event, 3 events per post. Number of listeners on the horizontal axis, time on the vertical axis](https://github.com/user-attachments/assets/23d89186-ba85-428a-9e70-ab61b9a59f2f)

The biggest wins come from when you have only a handful of listeners at most. The zero overhead for posting unused events
is especially nice.

<details>
<summary>Sidenote</summary>

> Why are the posting benchmark numbers so different from the December preview shown on the Discord?

This is to do with the handling of duplicate listeners. In the December preview, listeners were not deduplicated by
NovaBus, and I confirmed the number of listeners was as expected, therefore, I didn't use the ClassFactory as I thought
it was unnecessary. This was further confirmed by the SubscriberLambda posting benchmarks performing similarly without
the ClassFactory for EventBus v6 and the remark/comment I remember seeing somewhere about how lambda listeners are not
deduplicated.

After further investigation, I found that while the number of listeners was indeed correct and not deduplicated, the
JVM was smart enough to tell that they all referred to the same static method and performed additional optimisations.
Note that these optimisations were not on NovaBus' end and explicitly deduplicating still provided much better results.

So technically, both the December preview and the current version have accurate benchmark numbers, but the December
preview was less representative of real-world performance. The December preview benchmarks acted as if each mod was
doing `addListener(LibraryMod::onEvent)` instead of actually unique methods specific to each mod
(like `addListener(ModA::onEvent)`, `addListener(ModB::onEvent)`, etc...). The current version uses unique methods for
each listener.

The numbers are fair in both versions because I did like-for-like comparisons with Forge EventBus without any explicit
deduplication. In other words, the December preview benchmarks also skipped the ClassFactory for Forge EventBus and
both used the SubscriberLambda suite, so the numbers were still comparable. I opted to use the ClassFactory again for
this PR's numbers for more realistic results.
</details>

#### Posting data
<details>
<summary>(Click to show/hide)</summary>

```
EventBus v6 benchmark                                    Mode  Cnt     Score     Error  Units
BenchmarkModLauncher.Posting.Dynamic.postDynamicSingle   avgt    5    17.236 ±   0.111  ns/op
BenchmarkModLauncher.Posting.Dynamic.postDynamicDozen    avgt    5   156.574 ±  11.053  ns/op
BenchmarkModLauncher.Posting.Dynamic.postDynamicHundred  avgt    5  4281.470 ± 522.737  ns/op

BenchmarkModLauncher.Posting.Lambda.postLambdaSingle     avgt    5    17.118 ±   0.237  ns/op
…