@Jonathing 2025-03-12
After spending a not-insignificant amount of time notating the JavaDocs for the API, I feel the need to further express just how impressed I am by all of this. Not only is the API incredibly concise, but even the internal implementations are concise enough. Even though it's kind of hard to read (again, unavoidable considering we're squeezing the JVM for money at this point), it's still clear what each piece of the implementation is meant to do. And this is even more-so for the API.

I think in terms of the API, the only comments I have are to potentially re-assess `MutableEvent`. I think it was discussed here earlier but since it's only designed to work with `MonitorAware`, simply calling it `MonitorAwareEvent` would be enough. That is, unless, there are plans to expand on the capabilities of `MutableEvent`, but from the looks of it I don't see that happening. Other than that, it's just a matter of cleaning up and making sure everything is as clear as possible to both the maintainers and the consumers. I'm sure Lex has been keeping you company with implementation details anyways.