@PaintNinja 2025-03-17
> After spending a not-insignificant amount of time notating the JavaDocs for the API, I feel the need to further express just how impressed I am by all of this. Not only is the API incredibly concise, but even the internal implementations are concise enough -- even though it's kind of hard to read (again, unavoidable considering we're squeezing the JVM for money at this point), it's still clear what each piece of the implementation is meant to do. And this is even more-so for the API.

Thank you, I put a lot of thought into the API design and minimising complexity, I'm glad it's paid off. :)

> I think in terms of the API, the only comments I have are to potentially re-assess `MutableEvent`. I think it was discussed here earlier but since it's only designed to work with MonitorAware, simply calling it MonitorAwareEvent would be enough.

I've explained it here, I think there was a misunderstanding of its purpose that should be solved with better Javadocs in future: https://github.com/MinecraftForge/EventBus/pull/69#discussion_r1989659011

> There are a couple of Gradle-related issues with your tests that were reported in the problems file. We need to make sure to account for them.

These were already broken before the PR, but I've gone ahead and fixed them anyway.