@Jonathing 2025-03-17
> Thank you, I put a lot of thought into the API design and minimising complexity, I'm glad it's paid off. :)

Yes! And I hope our recent conversations in voice chat have helped to solidify this. A lot of effort has been put on all three of our parts into making a better understanding of our projects while also focusing scope. I'm very grateful for that.

While it's probably biased based off of my personal experiences with other projects and teams, I think part of the growing pain with this review has been to see how far we can get within scope of EventBus's usage. It's good to, at the bare minimum, try to figure out what we can do within the new EventBus to see if it'd be a reasonable thing to expect of consumers in the future. And when it's not worth it, detailing the reason for that should be given as much care as detailing *why* a new feature should be added, something I've meticulously done within the PRs I've triaged for Forge itself.

> I've explained it here, I think there was a misunderstanding of its purpose that should be solved with better Javadocs in future.

Yes, I agree. I think the JavaDocs will also need to be expanded upon over time as we get more comfortable with the redesign of EB7. Aside from the fact that we (in general, not specifically) are historically bad at naming things, keeping things well documented with use cases, usages, and abilities is going to keep the maintenance of the API in a good state.